Separate youtube ingestor
